Hex print bugfix. Formatted output.

+++ src/arch/x86_64/lib.h
@@ -1,7 +1,9 @@
 #ifndef __LIB_H
 #define __LIB_H
 
+#include "types.h"
+
 extern void memcpy(void* dest, void* src, int numOfBytes);
-extern char* itoa(int num, char* string, int base);
+extern char* itoa(uint64_t num, char* string, int base);
 
 #endif
